[FIX] quality_control_oca: inherit from mail tracking in qc_trigger_line
