+++ quality_control_oca/models/__init__.py
@@ -8,3 +8,6 @@ from . import qc_inspection
 from . import product_product
 from . import product_template
 from . import product_category
+from . import qc_trigger_product_category_line
+from . import qc_trigger_product_line
+from . import qc_trigger_product_template_line
